adjust login page for mobile
