+++ frontend/src/lib/styles/default.scss
@@ -1,4 +1,4 @@
-/* src/lib/styles/default.css */
+/* src/lib/styles/default.scss */
 
 @import './reset.scss';
