+++ frontend/src/lib/styles/mixins.scss
@@ -1,3 +1,5 @@
+/* src/lib/styles/mixins.scss */
+
 @mixin sp {
   @media screen and (max-width: 767px) {
     @content;
